requirements.txt
python-telegram-bot==20.7
ffmpeg-python==0.2.0
Pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
psutil==5.9.6 